requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
flask-limiter==3.5.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
requests==2.31.0
python-dotenv==1.0.0
pyjwt==2.8.0
authlib==1.2.1
pygments==2.17.2
gitpython==3.1.40
pandas==2.1.4
matplotlib==3.8.2
plotly==5.18.0
fpdf2==2.7.6
celery==5.3.6
redis==5.0.1
marshmallow==3.20.1
flask-marshmallow==1.2.0
marshmallow-sqlalchemy==1.0.0
click==8.1.7
rich==13.7.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.12.0
flake8==7.0.0
gunicorn==21.2.0
werkzeug==3.0.1
cryptography==41.0.7
httpx==0.26.0
tenacity==8.2.3
cachetools==5.3.2
python-dateutil==2.8.2
pytz==2023.3.post1
humanize==4.9.0
